ml_pipeline/requirements.txt
torch>=2.0.0
transformers>=4.30.0
firebase-admin>=6.2.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
python-dotenv>=1.0.0 